app/requirements.txt
Django>=4.0,<5.0
psycopg2-binary>=2.8
pytest==5.4.1
pytest-django==3.8.0
djangorestframework==3.12.2
django-rest-swagger==2.2.0
black==20.8b1
isort==5.6.4
autoflake==1.4
django-cors-headers==3.6.0
gunicorn==20.1.0
pytz==2021.3
tzdata==2021.5
qrcode==7.3.1
Pillow==9.0.1
geopy==2.2.0
drf-yasg==1.20.0